Refactor file writing to its own module and convert tyepscript
